pyproject.toml
[project]
name = "what-if-simulator"
version = "0.1.0"
description = "A multi-agent daily life consequence simulator built with LangChain and LangGraph."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
what-if-simulator = "simulation.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/simulation", "src/agents"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
